+++ packages/client/hmi-client/src/components/model/petrinet/model-configurations/tera-stratified-model-configuration.vue
@@ -15,18 +15,13 @@
 			<td
 				class="p-frozen-column second-frozen"
 				tabindex="0"
-				@keyup.enter="
-					modelConfigInputValue = cloneDeep(modelConfigurations[i].name)
-					// cellEditStates[i].name = true;
-				"
-				@click="
-					modelConfigInputValue = cloneDeep(modelConfigurations[i].name)
-					// cellEditStates[i].name = true;
-				"
+				@keyup.enter="emit('enter-name-cell', i)"
+				@click="emit('enter-name-cell', i)"
 			>
 				<InputText
-					v-if="cellEditStates[i]?.name"
-					v-model.lazy="modelConfigInputValue"
+					v-if="cellEditStates[i].name"
+					:value="editValue"
+					@input="emit('update:editValue', ($event.target as HTMLInputElement).value)"
 					v-focus
 					@focusout="emit('update-name', i)"
 					@keyup.stop.enter="emit('update-name', i)"
@@ -36,8 +31,8 @@
 					{{ name }}
 				</span>
 			</td>
-			<td v-for="(id, j) in [...baseStates, ...baseTransitions]" :key="j">
-				<section class="editable-cell" @click="openValueModal(id, i)">
+			<td v-for="(id, j) in baseStatesAndTransitions" :key="j">
+				<section class="editable-cell" @click="emit('open-matrix-modal', i, id)">
 					<span>{{ id }}<i class="pi pi-table" /></span>
 					<Button
 						class="p-button-icon-only p-button-text p-button-rounded p-button-icon-only-small cell-menu"
@@ -47,137 +42,39 @@
 			</td>
 		</tr>
 	</tbody>
-	<Teleport to="body">
-		<tera-modal
-			v-if="openValueConfig"
-			@modal-mask-clicked="openValueConfig = false"
-			@model-submit="openValueConfig = false"
-		>
-			<template #header>
-				<h4>{{ modalVal.id }}</h4>
-				<span>Configure the matrix values</span>
-				<div class="flex align-items-center">
-					<Checkbox
-						inputId="matrixShouldEval"
-						v-model="matrixShouldEval"
-						:binary="true"
-						label="Evaluate expressions?"
-					/>
-					<label for="matrixShouldEval" class="ml-2"> Evaluate Expressions? </label>
-				</div>
-			</template>
-			<template #default>
-				<!-- TODO: Implement value tabs for the modal once we are ready
-					<TabView v-model:activeIndex="activeIndex">
-					<TabPanel v-for="(extraction, i) in extractions" :key="i">
-						<template #header>
-							<span>{{ extraction.name }}</span>
-						</template>
-						<div>
-							<label for="name">Name</label>
-							<InputText class="p-inputtext-sm" :key="'name' + i" v-model="extraction.name" />
-						</div>
-						<div>
-							<label for="name">Matrix</label>
-							<tera-stratified-value-matrix
-								:model-configuration="modelConfigurations[modalVal.configIndex]"
-								:id="modalVal.id"
-								:stratified-model-type="stratifiedModelType"
-								:node-type="modalVal.nodeType"
-							/>
-						</div>
-					</TabPanel>
-				</TabView> -->
-				<tera-stratified-value-matrix
-					:model-configuration="modelConfigurations[modalVal.configIndex]"
-					:id="modalVal.id"
-					:stratified-model-type="stratifiedModelType"
-					:node-type="modalVal.nodeType"
-					:should-eval="matrixShouldEval"
-				/>
-			</template>
-			<template #footer>
-				<Button label="OK" @click="openValueConfig = false" />
-				<Button class="p-button-outlined" label="Cancel" @click="openValueConfig = false" />
-			</template>
-		</tera-modal>
-	</Teleport>
+	<Teleport to="body"> </Teleport>
 </template>
 
 <script setup lang="ts">
-import { ref } from 'vue';
-import { cloneDeep } from 'lodash';
 import Button from 'primevue/button';
-import TeraModal from '@/components/widgets/tera-modal.vue';
-// import TabPanel from 'primevue/tabpanel';
-// import TabView from 'primevue/tabview';
 import InputText from 'primevue/inputtext';
-import Checkbox from 'primevue/checkbox';
 import { ModelConfiguration } from '@/types/Types';
-import TeraStratifiedValueMatrix from '@/components/models/tera-stratified-value-matrix.vue';
-import { NodeType } from '@/model-representation/petrinet/petrinet-renderer';
-import { StratifiedModelType } from '@/model-representation/petrinet/petrinet-service';
-import { FeatureConfig } from '@/types/common';
 
-const props = defineProps<{
-	featureConfig: FeatureConfig;
+defineProps<{
+	editValue: string;
 	modelConfigurations: ModelConfiguration[];
-	stratifiedModelType: StratifiedModelType;
 	calibrationConfig?: boolean;
-	tableHeaders: { name: string; colspan: number }[];
 	cellEditStates: any[];
-	baseStates: any;
-	baseTransitions: any;
+	baseStatesAndTransitions: any;
 }>();
 
 const emit = defineEmits([
 	'new-model-configuration',
 	'update-name',
 	'update-value',
-	'enter-value-cell'
+	'enter-name-cell',
+	'enter-value-cell',
+	'open-matrix-modal',
+	'update:editValue'
 ]);
-
-const modelConfigInputValue = ref<string>('');
-// const extractions = ref<any[]>([]);
-const openValueConfig = ref(false);
-const modalVal = ref({ id: '', configIndex: 0, nodeType: NodeType.State });
-const matrixShouldEval = ref(true);
-
-const activeIndex = ref(0);
 
 // Makes cell inputs focus once they appear
 const vFocus = {
 	mounted: (el) => el.focus()
 };
-
-function openValueModal(id: string, configIndex: number) {
-	if (!props.featureConfig.isPreview) {
-		const nodeType = props.baseStates.includes(id) ? NodeType.State : NodeType.Transition;
-
-		activeIndex.value = 0;
-		openValueConfig.value = true;
-		modalVal.value = { id, configIndex, nodeType };
-	}
-}
 </script>
 
 <style scoped>
-.model-configuration:deep(.p-column-header-content) {
-	color: var(--text-color-subdued);
-}
-
-/** TODO: Apply to all tables in theme or create second table rules?  */
-.p-datatable-thead th {
-	text-transform: none !important;
-	color: var(--text-color-primary) !important;
-	font-size: var(--font-size-small) !important;
-	padding-left: 1rem !important;
-}
-
-.model-configuration:deep(.p-datatable-tbody > tr > td:empty:before) {
-	content: '--';
-}
-
 .cell-menu {
 	visibility: hidden;
 }
