+++ packages/client/hmi-client/src/types/common.ts
@@ -19,6 +19,12 @@ export interface FeatureConfig {
 	isPreview: boolean;
 }
 
+export enum ParamType {
+	CONSTANT = 'constant',
+	DISTRIBUTION = 'distribution',
+	TIME_SERIES = 'time_series'
+}
+
 // TODO: Wherever these are used - investigate using an actual map instead, this has been avoided due to v-model not playing well with maps
 // But a solution might be found here: https://stackoverflow.com/questions/37130105/does-vue-support-reactivity-on-map-and-set-data-types/64512468#64512468
 export interface StringValueMap {
